Rename the sf_calls file
